services/file-processor/requirements.txt
# 数据源服务依赖包
# 历史文本项目

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端和爬虫
aiohttp==3.9.1
httpx==0.25.2
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# 数据库
asyncpg==0.29.0
motor==3.3.2  # MongoDB异步驱动
pymongo==4.6.0

# 缓存
redis==5.0.1

# 文件存储
minio==7.2.0

# 消息队列
aio-pika==9.3.1  # RabbitMQ异步客户端

# 代理管理
aiohttp-socks==0.8.4
pysocks==1.7.1

# 基础图像处理（保留轻量级）
pillow==10.1.0

# PDF处理
PyPDF2==3.0.1
pdfplumber==0.10.3

# Word文档处理
python-docx==0.8.11

# 文本编码检测
chardet==5.2.0

# 可选依赖：根据需要启用
# pandas==2.1.3           # 数据分析 (93MB)
# numpy==1.25.2           # 数值计算 (38MB) 
# celery==5.3.4           # 异步任务队列
# kombu==5.3.4            # 消息中间件
# ddddocr==1.4.11         # 验证码识别 (85MB)

# 工具类
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3  # 重试机制

# 监控
prometheus-client==0.19.0
psutil==5.9.6

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0